// ExistingCode/src/data-access/domain/tables/assaybatchstatus/schema.ts
export const AssayBatchStatusTableInsertBaseSchema = z.object({
	/**
	 * dataType: uniqueidentifier
	 * defaultValue: (newid())
	 */
	assayBatchStatusId: z.any().optional(),
	/**
	 * dataType: nvarchar
	 */
	code: z.string().max(50),
	/**
	 * dataType: nvarchar
	 */
	description: z.string().max(60).nullable().optional(),
	/**
	 * dataType: smallint
	 * defaultValue: ((0))
	 */
	sortOrder: z.number().int().optional(),
	/**
	 * dataType: bit
	 * defaultValue: ((0))
	 */
	isDefaultInd: z.any().optional(),
	/**
	 * dataType: bit
	 * defaultValue: ((1))
	 */
	activeInd: z.any().optional(),
	/**
	 * dataType: datetimeoffset
	 */
	createdOnDt: z.any(),
	/**
	 * dataType: nvarchar
	 * defaultValue: (suser_sname())
	 */
	createdBy: z.string().max(50).optional(),
	/**
	 * dataType: datetimeoffset
	 */
	modifiedOnDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	modifiedBy: z.string().max(50).nullable().optional(),
	/**
	 * dataType: timestamp
	 */
	rv: z.date(),
});